+++ packages/ui/src/form/primitives/fieldset/details.tsx
@@ -2,8 +2,9 @@ import clsx from 'clsx';
 import { forwardRef, useEffect, useId } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import fieldStyles from '../../../utils/css/field.module.css';
-import { useFieldsetContext } from './context';
 import { VisuallyHidden } from '../../../visually-hidden';
+import { useFieldsetContext } from './context';
+import styles from './style.module.css';
 import type { FieldsetDetailsProps } from './types';
 
 /**
@@ -21,7 +22,7 @@ import type { FieldsetDetailsProps } from './types';
 export const FieldsetDetails = forwardRef<
 	HTMLDivElement,
 	FieldsetDetailsProps
->( function FieldsetDetails( { className, ...restProps }, ref ) {
+>( function FieldsetDetails( { className, children, ...restProps }, ref ) {
 	const id = useId();
 	const { registerDescriptionId, unregisterDescriptionId } =
 		useFieldsetContext( 'Fieldset.Details' );
@@ -32,15 +33,19 @@ export const FieldsetDetails = forwardRef<
 	}, [ registerDescriptionId, unregisterDescriptionId, id ] );
 
 	return (
-		<>
+		<div
+			ref={ ref }
+			className={ clsx(
+				styles.description,
+				fieldStyles.description,
+				className
+			) }
+			{ ...restProps }
+		>
 			<VisuallyHidden id={ id }>
 				{ __( 'More details follow.' ) }
 			</VisuallyHidden>
-			<div
-				ref={ ref }
-				className={ clsx( fieldStyles.description, className ) }
-				{ ...restProps }
-			/>
-		</>
+			{ children }
+		</div>
 	);
 } );